requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.8.0
